+++ src/module.ts
@@ -1,9 +1,10 @@
 import { resolve } from 'path'
 import { fileURLToPath } from 'url'
-import { defineNuxtModule, addPlugin } from '@nuxt/kit'
+import { defineNuxtModule, addPlugin, addComponentsDir } from '@nuxt/kit'
 import type { MetaData } from 'metadata-scraper/lib/types'
 import defu from 'defu'
 import { createPageGenerateHook, createGenerateDoneHook, CrawlerPage, CrawlerHooks } from './hooks'
+import type { DocSearchOptions } from './types'
 
 enum InstantSearchThemes {
   'reset',
@@ -17,6 +18,7 @@ interface ModuleBaseOptions {
   lite?: boolean;
   instantSearch?: boolean | { theme: keyof typeof InstantSearchThemes };
   recommend?: boolean;
+  docSearch?: Partial<DocSearchOptions>;
 }
 
 declare module '@nuxt/schema' {
@@ -44,7 +46,7 @@ export default defineNuxtModule<ModuleOptions>({
   meta: {
     name: '@nuxtjs/algolia',
     configKey: 'algolia',
-	compatibility: {
+    compatibility: {
       nuxt: '^3.0.0 || ^2.16.0',
       bridge: true
     }
@@ -54,6 +56,7 @@ export default defineNuxtModule<ModuleOptions>({
     apiKey: '',
     lite: true,
     instantSearch: false,
+    docSearch: {},
     crawler: {
       apiKey: '',
       indexName: '',
@@ -62,6 +65,9 @@ export default defineNuxtModule<ModuleOptions>({
     }
   },
   setup (options, nuxt) {
+    const runtimeDir = fileURLToPath(new URL('./runtime', import.meta.url))
+    nuxt.options.build.transpile.push(runtimeDir)
+
     if (!options.apiKey) {
       throw new Error('`[@nuxtjs/algolia]` Missing `apiKey`')
     }
@@ -87,11 +93,28 @@ export default defineNuxtModule<ModuleOptions>({
       })
     }
 
-    nuxt.options.publicRuntimeConfig.algolia = defu(nuxt.options.publicRuntimeConfig.algolia, {
+    if (Object.keys(options.docSearch).length) {
+      const docSearchConfig = options.docSearch
+
+      // Defaults apiKey and applicationId to global Algolia keys if not specified by the user
+      if (!docSearchConfig.apiKey && options.apiKey) { docSearchConfig.apiKey = options.apiKey }
+      if (!docSearchConfig.applicationId && options.applicationId) { docSearchConfig.applicationId = options.applicationId }
+
+      addComponentsDir({
+        path: resolve(runtimeDir, 'components'),
+        pathPrefix: false,
+        prefix: '',
+        level: 999,
+        global: true
+      })
+    }
+
+    nuxt.options.runtimeConfig.public.algolia = defu(nuxt.options.runtimeConfig.public.algolia, {
       apiKey: options.apiKey,
       applicationId: options.applicationId,
       lite: options.lite,
       instantSearch: options.instantSearch,
+      docSearch: options.docSearch,
       recommend: options.recommend
     })
 
@@ -110,8 +133,6 @@ export default defineNuxtModule<ModuleOptions>({
       }
     }
 
-    const runtimeDir = fileURLToPath(new URL('./runtime', import.meta.url))
-    nuxt.options.build.transpile.push(runtimeDir)
     addPlugin(resolve(runtimeDir, 'plugin'))
 
     nuxt.hook('autoImports:dirs', (dirs) => {